Read locale from URL pathname
This does come with some downsides. In some places the locale may need
to be prop-drilled down to where it is used. Also we give up on any
useful pathname typings.

It might be interesting to switch to
[next-intl](https://next-intl-docs.vercel.app/docs/getting-started/app-router/with-i18n-routing).
It appears to solve the locale prop-drilling and static typing, but I'm
not crazy about [ICU message
format](https://unicode-org.github.io/icu/userguide/format_parse/messages/).
I like that it's a standard, but it doesn't seem as ergonomic as
`i18next`'s format. Maybe I'm just used to what I know. I could be
convinced to switch if someone cares enough.